viz/requirements.txt
streamlit==1.32.0
trino==0.327.0
pandas==2.1.0
plotly==5.18.0
python-dotenv==1.0.0
